# openapi.yaml
openapi: 3.0.0
info:
  title: 小蚁搬运 SaaS 平台 API
  description: 小蚁搬运是一个SaaS架构的多租户跑腿装卸平台，支持货物的装卸搬运工作。
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: 开发环境
  - url: https://api.xiaoyibanyun.com/api
    description: 生产环境
paths:
...
  /admin/referral/campaigns/{id}/end:
    put:
      summary: 结束推荐活动（管理后台）
      tags:
        - AdminReferral
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: x-tenant-code
          required: true
          schema:
            type: string
          description: 租户编码
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 活动ID
      responses:
        '200':
          description: 结束成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "推荐活动已结束"
        '401':
          description: 未授权
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 活动不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/payments/wechat/notify:
    post:
      summary: 微信支付回调通知
      description: 微信支付结果通知回调接口
      tags:
        - Payments
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: object
              description: 微信支付回调数据（XML格式）
      responses:
        '200':
          description: 处理成功
          content:
            application/xml:
              schema:
                type: string
                example: "<xml><return_code>SUCCESS</return_code><return_msg>OK</return_msg></xml>"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: 第三方平台 API Key
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: "错误信息"

    # ======================
    # 第三方订单相关 Schema
    # ======================
...